pyproject.toml
[project]
name = "openhost_test_harness"
version = "0.1.0"
description = "Reusable test scaffolding for Openhost apps (mock router, etc.)"
requires-python = ">=3.12"
dependencies = [
    "attrs>=23",
    "httpx>=0.27",
    "websockets>=12",
    "hypercorn>=0.17",
    "litestar>=2.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openhost_test_harness"]
